requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
supabase==2.3.0
python-dotenv==1.0.0